pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitpatrotator"
version = "0.1.0"
description = "Cross-platform tool for rotating GitHub/GitLab PATs stored in HashiCorp Vault"
authors = [{name = "GitPATRotator Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Security",
]
dependencies = [
    "hvac>=1.2.1",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "PyJWT>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
gitpatrotator = "gitpatrotator.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/gitpatrotator"
Issues = "https://github.com/yourusername/gitpatrotator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["gitpatrotator*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=gitpatrotator --cov-report=term-missing"